Move Bookmark feature from exam to course
