+++ exam/src/main/AndroidManifest.xml
@@ -63,11 +63,6 @@
             android:theme="@style/TestpressTheme"
             android:configChanges="orientation|keyboard|screenSize" />
 
-        <activity android:name=".ui.BookmarksActivity"
-            android:label="@string/testpress_bookmarks"
-            android:theme="@style/TestpressTheme"
-            android:configChanges="orientation|keyboard|screenSize" />
-
         <activity android:name=".ui.ShareToUnLockActivity"
             android:label="@string/testpress_share"
             android:theme="@style/TestpressTheme"
